+++ src/rxify.ts
@@ -2,7 +2,7 @@ import { ApolloClient, ApolloQueryResult, ObservableQuery } from 'apollo-client'
 import { observeOn } from 'rxjs/operator/observeOn';
 import { AsyncAction } from 'rxjs/scheduler/AsyncAction';
 import { AsyncScheduler } from 'rxjs/scheduler/AsyncScheduler';
-import { isFunction } from 'lodash';
+import { isObject } from './utils/helpers';
 
 import { createWithObservableVariables } from './utils/variables';
 import { RxObservableQuery } from './RxObservableQuery';
@@ -25,7 +25,7 @@ export function rxify(
   }
 
   // ApolloClient.watchQuery
-  if (isFunction(clientOrWatchQuery)) {
+  if (isObject(clientOrWatchQuery)) {
     return wrapWatchQuery(clientOrWatchQuery);
   }